% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreedyExperimentalDesign.R
\docType{package}
\name{GreedyExperimentalDesign}
\alias{GreedyExperimentalDesign}
\alias{GreedyExperimentalDesign-package}
\title{Greedy Experimental Design Search}
\description{
A tool to find a priori experimental designs with good balance greedily
}
\references{
Kapelner, A
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{Design,}
\keyword{Experimental}
\keyword{Greedy}
\keyword{search}
