% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_numerical_optimization_search.R
\name{gurobi_multiple_designs}
\alias{gurobi_multiple_designs}
\title{Find multiple designs}
\usage{
gurobi_multiple_designs(X, r, ...)
}
\arguments{
\item{X}{The design matrix with $n$ rows (one for each subject) and $p$ columns 
(one for each measurement on the subject). This is the design matrix you wish 
to search for a more optimal design.}

\item{r}{The number of vectors that should be returned}

\item{...}{Additional arguments to be passed to \code{initGurobiNumericalOptimizationExperimentalDesignObject}.}
}
\value{
A matrix of allocation vectors of dimension \code{r x n}.
}
\description{
This method searches through $1_{T}$ space using Gurobi's optimization many times.
It finds many different solutions by permuting the rows of the design matrix and 
rerunning the optimization.
}
\author{
Kapelner
}
