% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_search.R
\name{resultsOptimalSearch}
\alias{resultsOptimalSearch}
\title{Returns the results (thus far) of the optimal design search}
\usage{
resultsOptimalSearch(obj)
}
\arguments{
\item{obj}{The \code{optimal_experimental_design} object that is currently running the search}
}
\description{
Returns the results (thus far) of the optimal design search
}
\author{
Adam Kapelner
}
